backend/requirements.txt
asyncio==3.4.3
bidict==0.21.2
python-engineio==4.0.1
python-socketio==5.1.0
tornado==6.1
